Move the page files out of their folders
