+++ src/App.js
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { Routes, Route } from 'react-router-dom';
-import RocketsPage from './pages/rockets/RocketsPage';
+import RocketsPage from './pages/RocketsPage';
 import { fetchMissionsData } from './redux/missions/missionsSlice';
-import MissionsPage from './pages/missions/MissionsPage';
-import ProfilePage from './pages/profile/ProfilePage';
+import MissionsPage from './pages/MissionsPage';
+import ProfilePage from './pages/ProfilePage';
 import Layout from './components/layout/Layout';
 import './App.css';
 import { fetchRocketsData } from './redux/rockets/rocketsSlice';
